move load_metadata_standard
